# crates/veronex/src/infrastructure/inbound/http/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Veronex API",
    "description": "Inference gateway API — LLM routing, job tracking, API key management, and cluster observability.",
    "version": "0.1.0",
    "contact": {
      "name": "Veronex",
      "url": "https://veronex.verobee.com"
    }
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "Local dev" }
  ],
  "security": [{ "ApiKeyAuth": [] }],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
      "JobStatus": {
        "type": "string",
        "enum": ["pending", "running", "completed", "failed", "cancelled"]
      },
      "ProviderType": {
        "type": "string",
        "enum": ["llama_server", "gemini"]
      },
      "SubmitInferenceRequest": {
        "type": "object",
        "required": ["model", "prompt"],
        "properties": {
          "model": { "type": "string", "example": "llama3.2:3b" },
          "prompt": { "type": "string", "example": "Explain hexagonal architecture." },
          "provider_type": { "$ref": "#/components/schemas/ProviderType" },
          "stream": { "type": "boolean", "default": false }
        }
      },
      "JobResponse": {
        "type": "object",
        "properties": {
          "job_id": { "type": "string", "format": "uuid" },
          "status": { "$ref": "#/components/schemas/JobStatus" },
          "model": { "type": "string" },
          "provider_type": { "$ref": "#/components/schemas/ProviderType" },
          "result": { "type": "string", "nullable": true },
          "error": { "type": "string", "nullable": true },
          "prompt_tokens": { "type": "integer", "nullable": true },
          "completion_tokens": { "type": "integer", "nullable": true },
          "latency_ms": { "type": "integer", "nullable": true },
          "ttft_ms": { "type": "integer", "nullable": true },
          "tps": { "type": "number", "nullable": true },
          "created_at": { "type": "string", "format": "date-time" },
          "started_at": { "type": "string", "format": "date-time", "nullable": true },
          "completed_at": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
...
      "OpenAIChatRequest": {
        "type": "object",
        "required": ["model", "messages"],
        "properties": {
          "model": { "type": "string", "example": "llama3.2:3b" },
          "messages": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "role": { "type": "string", "enum": ["system", "user", "assistant"] },
                "content": { "type": "string" }
              }
            }
          },
          "stream": { "type": "boolean", "default": false },
          "temperature": { "type": "number" },
          "max_tokens": { "type": "integer" }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      }
    }
  },
...
  "paths": {
...
    "/v1/inference": {
      "post": {
        "tags": ["inference"],
        "summary": "Submit inference job",
        "operationId": "submitInference",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SubmitInferenceRequest" }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Job accepted",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JobResponse" } } }
          },
          "400": { "description": "Bad request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "Unauthorized" },
          "429": { "description": "Rate limit exceeded" }
        }
      }
    },
    "/v1/inference/{job_id}/stream": {
      "get": {
        "tags": ["inference"],
        "summary": "Stream inference output (SSE)",
        "operationId": "streamInference",
        "parameters": [
          { "name": "job_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "SSE stream of inference tokens",
            "content": { "text/event-stream": { "schema": { "type": "string" } } }
          },
          "404": { "description": "Job not found" }
        }
      }
    },
    "/v1/inference/{job_id}/status": {
      "get": {
        "tags": ["inference"],
        "summary": "Get job status",
        "operationId": "getJobStatus",
        "parameters": [
          { "name": "job_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Job status",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JobResponse" } } }
          },
          "404": { "description": "Job not found" }
        }
      }
    },
    "/v1/inference/{job_id}": {
      "delete": {
        "tags": ["inference"],
        "summary": "Cancel inference job",
        "operationId": "cancelInference",
        "parameters": [
          { "name": "job_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Job cancelled" },
          "404": { "description": "Job not found" }
        }
      }
    },
    "/v1/chat/completions": {
      "post": {
        "tags": ["openai"],
        "summary": "OpenAI-compatible chat completions",
        "description": "Drop-in replacement for OpenAI's /v1/chat/completions. Supports streaming via SSE.",
        "operationId": "chatCompletions",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/OpenAIChatRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat completion response or SSE stream",
            "content": {
              "application/json": { "schema": { "type": "object" } },
              "text/event-stream": { "schema": { "type": "string" } }
            }
          },
          "401": { "description": "Unauthorized" },
          "429": { "description": "Rate limit exceeded" }
        }
      }
    },
...
  }
}